chore: move everything to internal package
