+++ internal/config/testdata/json/default.json
@@ -1,0 +1,3 @@
+{
+  
+}
