--- pkg/config/testdata/git/custom/glu.yml
@@ -1,10 +1,0 @@
-sources:
-  git:
-    custom:
-      remote:
-        name: origin
-        url: https://corp-repos/custom
-        credential: vault
-        interval: 1m
-      path: v1
-      default_branch: release-v1
