--- pkg/config/testdata/git/default/glu.yaml
@@ -1,6 +1,0 @@
-sources:
-  git:
-    default:
-      remote:
-        name: upstream
-        url: https://corp-repos/default.git
