--- pkg/config/testdata/history/file/default/glu.yml
@@ -1,4 +1,0 @@
-history:
-  file:
-    default:
-      path: "history.db"
