--- pkg/config/testdata/json/glu.json
@@ -1,16 +1,0 @@
-{
-  "sources": {
-    "git": {
-      "custom": {
-        "remote": {
-          "name": "origin",
-          "url": "https://corp-repos/custom",
-          "credential": "vault",
-          "interval": "1m"
-        },
-        "path": "v1",
-        "default_branch": "release-v1"
-      }
-    }
-  }
-}
